pyproject.toml
[project]
name = "sachetai-pathway"
version = "0.1.0"
description = "NDRF disaster response system for Maharashtra using real-time RAG"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.4.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build. targets.wheel]
packages = ["api", "data_sources", "pipeline", "rag"]

[tool.hatch.build. targets.sdist]
include = [
    "*. py",
    "api/",
    "data_sources/",
    "pipeline/",
    "rag/",
]